Remove unused files and satart with Refactor the page section and Recrate it be comfortble with FSD pattern
