+++ frontend/src/(page)/(auth)/froget-password/page.tsx
