+++ frontend/src/(page)/(auth)/register/page.tsx
