+++ frontend/src/(page)/(main)/layout.tsx
