+++ frontend/src/(page)/(main)/product/[name]/page.tsx
