+++ frontend/src/app/hooks/auth/use-register.tsx
@@ -1,4 +1,4 @@
-import { RegisterCredentials } from "@/app/types/auth";
+import { RegisterCredentials } from "@/shared/types/auth";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { unknown } from "zod";
