+++ frontend/src/app/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { authService } from "../service/auth.service";
+import { authService } from "../../shared/service/auth.service";
 // import { USER_QK } from "../types/auth";
 
 const Authkey = {
