+++ frontend/src/shared/service/auth-server.ts
