+++ frontend/src/shared/service/auth.service.ts
