+++ frontend/src/shared/types/auth.d.ts
