+++ frontend/src/widgets/layouts/auth-layout/index.ts
@@ -1,0 +1,1 @@
+export * from './ui/auth-layout'
