+++ frontend/src/widgets/layouts/auth-layout/ui/auth-layout.tsx
@@ -1,0 +1,18 @@
+import ReturntoHome from "@/shared/ui/returnto-home";
+
+function AuthLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <div className="flex w-full min-h-screen bg-gray-50 dark:bg-gray-900 content-transition">
+            <ReturntoHome />
+            {children}
+        </div>
+    );
+}
+
+export {
+    AuthLayout
+}
