+++ frontend/src/widgets/layouts/index.ts
@@ -1,0 +1,2 @@
+export * from './main-layout'
+export * from './auth-layout'
