+++ frontend/src/widgets/layouts/main-layout/index.ts
@@ -1,0 +1,1 @@
+export * from './ui/main-layout'
