+++ frontend/src/widgets/layouts/main-layout/ui/main-layout.tsx
@@ -1,0 +1,28 @@
+import Sidbar from "@/widgets/sidebar/ui/sidebar";
+import { Providers } from "@/app/provider/theme-provider";
+import type { Metadata } from "next";
+
+
+export const metadata: Metadata = {
+    title: "Create Next App",
+    description: "Generated by create next app",
+};
+
+async function MainLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+
+    return (
+        <div className="flex  w-full min-h-screen bg-gray-50 dark:bg-gray-900 content-transition">
+            <Providers >
+                <Sidbar />
+                {children}
+            </Providers>
+        </div>
+    );
+}
+export {
+    MainLayout
+}
